Add files via upload
gap.py is the python based gap analysis tool.
CSV file is the result of findings in case of gaps (threshold is set as 0.5)
JSON file is with the ISO 27001 requirements - only a subset is given here. More requirements and sub requirements can be added to this file in the same format to get better results.